refactor: Remove file extensions from import paths for consistency
